# src/main/java/com/funnygorilla/tinybankcustomer/dto/CustomerDTO.java
package com.funnygorilla.tinybankcustomer.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.ToString;

@ToString
@Builder
@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class CustomerDTO {
	private Long id;
	private String firstName;
	private String lastName;
	private Boolean active;

}